# api/definition/swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://fingertips-next-api/1.0.0
info:
  description: An API to query public health indicator data.
  version: "1.0.0"
  title: Fingertips API
  contact:
    email: ProfileFeedback@dhsc.gov.uk
...
paths:
...
  /indicators:
    get:
      tags:
        - indicators
      summary: Get indicator summaries & filter results
      operationId: filterIndicators
      description: |
        Get summaries of public health indicators, by passing in the appropriate options you can filter the results
      parameters:
        - $ref: "#/components/parameters/indicator_ids"
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IndicatorSummary"
        "400":
          description: Bad request. The server could not process the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /indicators/{indicator_id}/data:
    get:
      tags:
        - indicators
      summary: Get health data for an indicator
      description:
        Get data for a public health indicator. This will return all data for all areas and all years for the indicators. Optionally filter the results by supplying one
        or more area codes and one or more years in the query string.
      responses:
        "200":
          description:
            Data containing the health data points for the indicator, years
            and geographical areas requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndicatorWithHealthDataForArea"

        "400":
          description: Bad request. The server could not process the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequest"
              examples:
                Too many area_codes:
                  value:
                    message: Too many values supplied for parameter area_codes. The maximum is 100 but 101 supplied.
                Too many years:
                  value:
                    message: Too many values supplied for parameter years. The maximum is 20 but 21 supplied.
        "500":
          $ref: "#/components/responses/InternalServerError"
      parameters:
        - $ref: "#/components/parameters/indicator_id"
        - $ref: "#/components/parameters/area_codes"
        - $ref: "#/components/parameters/area_type"
        - $ref: "#/components/parameters/years"
        - $ref: "#/components/parameters/inequalities"
      operationId: getHealthDataForAnIndicator
  /indicators/quartiles:
    get:
      tags:
        - indicators
      summary: Get quartile values for indicators
      description: |
        Get quartile information for indicators including 
      parameters:
        - $ref: "#/components/parameters/indicator_ids"
        - $ref: "#/components/parameters/area_code"
        - $ref: "#/components/parameters/area_type"
        - $ref: "#/components/parameters/ancestor_code"
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuartileData"
        "400":
          description: Bad request. The server could not process the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  schemas:
    IndicatorPolarity: # Define enum globally
      type: string
      enum: [ Unknown, NoJudgement, LowIsGood, HighIsGood ]
      x-enum-varnames: [ Unknown, NoJudgement, LowIsGood, HighIsGood ]
    BenchmarkComparisonMethod:
      type: string
      example: CIOverlappingReferenceValue95
      description: the benchmark comparison method used
      enum: [ Unknown, Quintiles, CIOverlappingReferenceValue95, CIOverlappingReferenceValue99_8 ]
      x-enum-varnames: [ Unknown, Quintiles, CIOverlappingReferenceValue95, CIOverlappingReferenceValue99_8 ]
    BenchmarkOutcome:
      type: string
      example: Lower
      description: benchmark outcome
      enum: [ NotCompared, Lowest, Lower, Low, Middle, High, Higher, Highest, Better, Best, Worse, Worst, Similar ]
      x-enum-varnames: [ NotCompared, Lowest, Lower, Low, Middle, High, Higher, Highest, Better, Best, Worse, Worst, Similar ]
...
    IndicatorSummary:
      type: object
      description: A summary of a public health indicator
      required:
        - indicator_id
        - title
      properties:
        indicator_id:
          type: integer
          format: int32
          example: 3456
          description: The unique identifier of the indicator
        title:
          type: string
          example: "Hypertension: QOF prevalence (all ages)"
          description: The title of the indicator
...
    HealthDataPoint:
      type: object
      description:
        Represents a health data point for a public health indicator with a
        count, value, upper confidence interval, lower confidence interval,
        year, age band and sex.
      required:
        - year
        - ageBand
        - sex
        - trend
        - deprivation
      properties:
        year:
          type: integer
          format: int32
          example: 2023
          description: The year that the data point is for
        count:
          type: number
          example: 222
          description: The count
        value:
          type: number
          example: 506.60912
          description: The value
        lowerCi:
          type: number
          example: 441.69151
          description: The lower confidence interval
        upperCi:
          type: number
          format: float
          example: 578.32766
          description: The upper confidence interval
        ageBand:
          $ref: "#/components/schemas/AgeData"
        sex:
          $ref: "#/components/schemas/SexData"
        trend:
          type: string
          example: Increasing
          description: The statistical trend that applies to the data point,
            given the preceding data. Will only be calculated if there are at
            least 5 data points to use.
          enum: [
            'Not yet calculated',
            'Cannot be calculated',
            'Increasing',
            'Decreasing',
            'No significant change',
            'Increasing and getting better',
            'Increasing and getting worse',
            'Decreasing and getting better',
            'Decreasing and getting worse'
          ]
        isAggregate:
          type: boolean
          description: Indicates if the datapoint is an aggregate point.
        benchmarkComparison:
          type: object
          properties:
            outcome:
              $ref: "#/components/schemas/BenchmarkOutcome"
            method:
              $ref: "#/components/schemas/BenchmarkComparisonMethod"
            indicatorPolarity:
              type: string
              example: HighIsGood
              description: the type of polarity applied to the outcome
              enum: [ NoJudgement, LowIsGood, HighIsGood ]
            benchmarkAreaCode:
              type: string
              example: E92000001
              description: The code of the benchmark area
            benchmarkAreaName:
              type: string
              example: England
              description: The name of the benchmark area
            benchmarkValue:
              type: number
              example: 123.5
              description: The value used in the comparison
        deprivation:
          $ref: "#/components/schemas/DeprivationData"
    DeprivationData:
      type: object
      description: Details of the deprivation category that this data point belongs to.
      required:
        - sequence
        - value
        - type
        - isAggregate
      properties:
        sequence:
          type: number
          format: integer
          example: 2
          description: A number which represents this deprivation value's inherent place in the sequence of all values for this deprivation type. Can be used to show values ordered by their deprivation dimension.
        value:
          type: string
          example: Second most deprived decile
          description: The deprivation value for this data point.
        type:
          type: string
          example: District & UA deprivation deciles in England (IMD2019, 4/21 geography)
          description: The deprivation category that this data point's deprivation value belongs to.
        isAggregate:
          type: boolean
          example: false
          description: Indicates if the datapoint is an aggregated value for the deprivation dimension.
    SexData:
      type: object
      description: Details of the sex category that this data point belongs to.
      required:
        - value
        - isAggregate
      properties:
        value:
          type: string
          example: Female
          description: Sex which the data are for.
        isAggregate:
          type: boolean
          example: false
          description: Indicates if the datapoint is an aggregated value for the sex dimension.
    AgeData:
      type: object
      description: Details of the age band that this data point belongs to.
      required:
        - value
        - isAggregate
      properties:
        value:
          type: string
          example: 0-4
          description: Age band which the data are for.
        isAggregate:
          type: boolean
          example: false
          description: Indicates if the datapoint is an aggregated value for the age dimension.
    HealthDataForArea:
      type: object
      description: Associates a list of health data points with the relevant
        geographical area (represented by it's unique code).
      required:
        - areaCode
        - areaName
        - healthData
      properties:
        areaCode:
          type: string
          example: E92000001
          description: The unique area code that the health data are for
        areaName:
          type: string
          example: England
          description: The name of the area that the health data are for
        healthData:
          type: array
          items:
            $ref: "#/components/schemas/HealthDataPoint"
          description: The health data points for the area and indicator
    IndicatorWithHealthDataForArea:
      type: object
      description: Indicator information with area health data
      properties:
        indicatorId:
          type: integer
          example: 21404
          description: Unique ID of the indicator
        name:
          type: string
          example: Emergency readmissions within 30 days of discharge from hospital
          description: Name of the indicator
        polarity:
          $ref: "#/components/schemas/IndicatorPolarity"
        benchmarkMethod:
          $ref: "#/components/schemas/BenchmarkComparisonMethod"
        areaHealthData:
          type: array
          items:
            $ref: "#/components/schemas/HealthDataForArea"
    QuartileData:
      type: object
      description: Quartile data for an indicator plus supporting information 
      properties:
        indicatorId:
          type: integer
          example: 21404
          description: Unique ID of the indicator
        year:
          type: integer
          format: int32
          example: 2023
          description: The year that the data point is for        
        polarity:
          $ref: "#/components/schemas/IndicatorPolarity"
        q0Value:
           type: number
           description: Quartile Q0 value - the minimum value
        q1Value:
           type: number
           description: Quartile Q1 value - the lower quartile value
        q2Value:
           type: number
           description: Quartile Q2 value - the median value
        q3Value:
           type: number
           description: Quartile Q3 value - the upper quartile value
        q4Value:
           type: number
           description: Quartile Q4 value - the maximum value
        areaValue:
          type: number
          description: the average value for the area requested for comparison 
        ancestorValue:
          type: number
          description: the average value for the ancestor area requested for comparison 
        englandValue:
          type: number
          description: the average value for England for comparison
    BadRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
      description:
        The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
  parameters:
    years:
      in: query
      name: years
      description: A list of years, up to 20 years can be requested
      style: form
      schema:
        type: array
        items:
          type: integer
          example: 2023
        maxItems: 20
    ancestor_code:
      in: path
      name: ancestor_code
      style: simple
      schema:
        type: string
        example: E12000001
      explode: false
      required: true
      description: The area code of an ancestor area
    area_code:
      in: path
      name: area_code
      style: simple
      schema:
        type: string
        example: G82109
      explode: false
      required: true
      description: The area code of the area/ geography
    inequalities:
      in: query
      name: inequalities
      style: form
      schema:
        type: array
        items:
          type: string
          enum:
            - age
            - sex
            - deprivation
      description: Determines the kind of inequality data that should be returned if an option is specified
...
    area_codes:
      in: query
      name: area_codes
      description: A list of area codes, up to 100 area codes can be requested
      style: form
      schema:
        type: array
        items:
          type: string
          example: G82109
        maxItems: 100
    area_type:
      in: query
      name: area_type
      description: The area type which the areas belong to
      schema:
        type: string
    indicator_id:
      in: path
      name: indicator_id
      style: simple
      schema:
        type: integer
        example: 1234
        minimum: 1
      explode: false
      required: true
      description: The unique identifier of the indicator
    indicator_ids:
      in: query
      name: indicator_ids
      description: A list of indicator_ids, up to 10 can be requested
      style: form
      schema:
        type: array
        items:
          type: integer
          example: 1234
        maxItems: 10
...
  responses:
...
    InternalServerError:
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
      description:
        The server encountered an unexpected condition that prevented it
        from fulfilling the request. Report the issue to the support team if it
        persists.